Agregadas Policies

+++ app/controllers/reports_controller.rb
@@ -5,20 +5,19 @@ class ReportsController < ApplicationController
     end
 
     def index
-        @reports = Report.with_attached_photo
-        pp Current.user::id
         if Current.user::admin==true then
             filter = Report::TIME_FILTER_ADMIN.fetch(params[:filter]&.to_sym, Report::TIME_FILTER['all'.to_sym])
+            @reports = Report.with_attached_photo.where(filter).order(created_at: :desc).load_async
         else
             filter = Report::TIME_FILTER.fetch(params[:filter]&.to_sym, Report::TIME_FILTER['all'.to_sym])
+            @reports = Report.with_attached_photo.where(filter).order(created_at: :desc).load_async
         end
-        pp "adasjdlkasjdasjdkjasljk"
-        @reports = @reports.where(filter).order(created_at: :desc).load_async
         @pagy, @reports = pagy_countless(@reports,items: 10)
 
     end
 
     def show
+        authorize report
         if report.visto==false then
             report.update(visto:true)
         end
@@ -38,10 +37,11 @@ class ReportsController < ApplicationController
 
     end
     def edit
-        report
+        authorize report
     end
 
     def update
+        authorize  report
         if report.update(report_params)
         
             redirect_to reports_path, notice: 'Reporte Actualizado'
@@ -52,6 +52,7 @@ class ReportsController < ApplicationController
     end
 
     def destroy
+        authorize report
         if report.destroy
             redirect_to reports_path,notice: 'Reporte Eliminado',status: :see_other
         else
